backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.2
Django==3.2.8
django-cors-headers==3.11.0
django-ipware==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
regex==2021.10.23
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
